requirement.txt
tensorflow==2.1.0
tensorflow_datasets==2.1.0
flask
py-healthcheck
gunicorn
google-cloud-logging